Wrapped GameQuitButton.java in abstract SimpleButton.java for more modular usage; Switched sizing logic to direct assignment
